% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpatial.R
\name{bottomright}
\alias{bottomright}
\title{Retrieve bottomright of \code{sp} object}
\usage{
bottomright(geom)
}
\arguments{
\item{geom}{Input geometry (polygon)}
}
\description{
For use with the (annoying) \code{openmap()} functions which retrieve basemap tiles for a box specified by the topleft and bottom right vertices.
}
\examples{
shade = gBoundingPoly(cambridge_town, tol = 1.1)
basemap = openmap(topleft(shade), bottomright(shade), type = "skobbler")

plot(basemap)
# skobbler maps look better with a little purple haze
plot(spTransform(shade, osm()), add = T, col = col2hex("lavender", "40"), border = NA)
plot(spTransform(cambridge_town, osm()), add = T, col = col2hex("skyblue", "80"))
}
\keyword{spatial}
