% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpatial.R
\name{col2hex}
\alias{col2hex}
\title{Convert R color strings to hex}
\usage{
col2hex(basecol, achannel = "")
}
\arguments{
\item{basecol}{An R base color name}

\item{achannel}{(ideally) a hex value string for the alpha channel. Decimal values will be interpreted as hex}
}
\description{
Quickly create a hex string for a color + alpha channel
}
\examples{
col2hex("lightsteelblue", "70")
}

