% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpatial.R
\name{gBoundingPoints}
\alias{gBoundingPoints}
\title{Select points within bounding box}
\usage{
gBoundingPoints(allpoints, b_box)
}
\arguments{
\item{allpoints}{A \code{SpatialPoints} or \code{SpatialPointsDataFrame} object}

\item{b_box}{A set of bounding coordinates in matrix form like those returned from \code{sp::bbox()}}
}
\description{
Uses bracket extraction to very quickly select points within a bounding box
}
\details{
Intersect methods (e.g. \code{rgeos::gContains} or \code{rgeos::gIntersection}) are very slow for extracting points via a bounding box. This function simply subsets a \code{SpatialPoints} or \code{SpatialPointsDataFrame} object using standard bracket extraction. As a result, it is much faster

\strong{Note}: all inputs should be of class \code{sp::SpatialPolygons|Points}
}
\examples{
gBoundingPoints(points.spdf, bbox(container_polygon))
}
\keyword{spatial}
