% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpatial.R
\name{gEnclosedCentroid}
\alias{gEnclosedCentroid}
\title{Return polygons whose centroids are enclosed}
\usage{
gEnclosedCentroid(container, reference)
}
\arguments{
\item{container}{The single polygon that will enclose the set of reference polygons}

\item{reference}{The set of polygons that will be selected down to the containing polygon}
}
\description{
We want to return the reference polygons whose centroids are contained within a larger, containing polygon
}
\details{
\strong{Note}: all inputs should be of class \code{sp::SpatialPolygons|Points}
}
\examples{
gEnclosedCentroid(drivetime, block_groups)
}
\keyword{spatial}

