% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpatial.R
\name{gPolyByCentroid}
\alias{gPolyByCentroid}
\title{Return polygons enclosed within a container}
\usage{
gPolyByCentroid(reference, centroids, container)
}
\arguments{
\item{reference}{A \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} object with the values to be summarized within a containing polygon}

\item{centroids}{A set of centroids for the reference polygons obtained by \code{gCentroids}}

\item{container}{A \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} object of a single, container polygon}
}
\description{
Pretty much a copy of \code{gEnclosedCentroid} but more robust
}
\details{
Returns reference polygons whose centroids are enclosed within a containing polygon. Allows for (requires) the pre-specification of reference polygon centroids which is useful when there are a very large number of reference polygons. Uses a bounding box filter before any spatial operations which makes it much faster than a straight intersection.

\strong{Note}: all inputs should be of class \code{sp::SpatialPolygons|Points}
}
\examples{
gPolyByCentroid(blockgroups, sp::gCentroid(blockgroups, byid = T), cbsa)
}
\keyword{spatial}

