% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpatial.R
\name{gWhichPoints}
\alias{gWhichPoints}
\title{Return indices of points within a box}
\usage{
gWhichPoints(allpoints, b_box)
}
\arguments{
\item{allpoints}{A \code{SpatialPoints} or \code{SpatialPointsDataFrame} object}

\item{b_box}{A set of bounding coordinates in matrix form like those returned from \code{sp::bbox()}}
}
\description{
Like \code{gBoundingPoints} but only returns the indices of points contained within a bounding box
}
\details{
Rather than returning the spatial object, this give the indices of points within a bounding box. It's more useful for re-merging attributes (Since gBoundingPoints strips a SpatialPointsDataFrame to a SpatialPoints object). Or, better still, subsetting.

\strong{Note}: all inputs should be of class \code{sp::SpatialPolygons|Points}
}
\examples{
gWhichPoints(points.spdf, bbox(container_polygon))
all_the_points_spdf[gWhichPoints(all_the_points.spdf, bbox(container_polygon)), ]
}
\keyword{spatial}
