% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpatial.R
\name{is.error}
\alias{is.error}
\title{Error Trap}
\usage{
is.error(x)
}
\arguments{
\item{any}{value returned from a function}
}
\description{
Copied from Hadley. Useful for providing a way for functions to exit gracefully from an error
}
\details{
Returns \code{TRUE} if \code{try} returns an error
}
\examples{
result = try(somefunction())
if(is.error(result)) {
print("Bailing out!")
} else {
print("result is not an error!")
}
}
\keyword{none}

