% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpatial.R
\name{mBox}
\alias{mBox}
\title{Return a square spatial polygon based on a radial distance from a point}
\usage{
mBox(x, y, size = 500, proj.srid = "+init=epsg:3857",
  gcs.srid = "+init=epsg:4326")
}
\arguments{
\item{size}{radial size of box given in units for the given projection (meters by default)}

\item{proj.srid}{a \code{proj4string} for a planar projected coordinate system. Default \code{epsg:3857} 'Web Mercator"}

\item{gcs.srid}{a \code{proj4string} for the input geographic (lat/lon) coordinate system and for the returned object. Default \code{epsg:4326} 'WGS84"}

\item{{x, }{y} Real values corresponding to longitude and latitude of a point}
}
\description{
Returns an object of type \code{SpatialPolygons} with a single feature.
}
\examples{
box_1k = mBox(-90, 30, size = 500)
}
\keyword{spatial}

