% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpatial.R
\name{mGetIntersecting}
\alias{mGetIntersecting}
\title{Return polygons enclosed within a container (Deprecated)}
\usage{
mGetIntersecting(container, reference)
}
\arguments{
\item{container}{A \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} object of a single, container polygon}

\item{reference}{A \code{SpatialPolygons} or \code{SpatialPolygonsDataFrame} object with the values to be summarized within a containing polygon}

\item{centroids}{A set of centroids for the reference polygons obtained by \code{gCentroids}}
}
\description{
Kinda like \code{gEnclosedArea}
}
\details{
Returns all the reference polygons from a SpatialPolygonsDataFrame that intersect with a container. Adds an attribute to the @data slot with the percent of overlapping area.

This function uses name matching / extraction. It's a kludge and often fails. use \code{gPolyByIntersect} instead. Preserved for existing code that uses it.

\strong{Note}: all inputs should be of class \code{sp::SpatialPolygons|Points}
}
\examples{
gPolyByCentroid(blockgroups, sp::gCentroid(blockgroups, byid = T), cbsa)
}
\keyword{spatial}

