% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpatial.R
\name{mSummarizer}
\alias{mSummarizer}
\title{summarize attributes over a containing geometry}
\usage{
mSummarizer(container, reference, buffer = 1, colnamevec, centroid = F,
  threshold = 0.99, FUN)
}
\arguments{
\item{container}{The thing you want to summarize within}

\item{reference}{The thing whose attributes you want to summarize}

\item{buffer}{How much you want to blow up the summarizing container (in meters)}

\item{colnamevec}{A vector of column names corresponding to those in the container whose values will be summarized}

\item{centroid}{Whether to use containing centroids for choosing reference geometries within the container}

\item{threshold}{the overlap percentage to use when \code{centroid = F}}
}
\description{
Convenience wrapper around gPolyByIntersect that applies a function over specified attributes in the reference set. It allows for buffering the input container.
}
\details{
\strong{Note}: careful, it's pretty fragile
}
\examples{
mSummarizer(cambridge_town, boston_blocks, buffer = 0, c("total_pop", "total_hh"), centroid = T, FUN = "sum")
}
\keyword{spatial}
