% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpatial.R
\name{osmGet}
\alias{osmGet}
\title{Plot OSM basemap}
\usage{
osmGet(geom, tolerance = 2, tileserver = "skobbler")
}
\arguments{
\item{geom}{Input geometry (polygon)}

\item{tolerance}{multiplier for the size of the bounding box around the input geometry}

\item{tileserver}{see the openmap() documentation for options}
}
\description{
Convenience wrapper around the openmap() function to return the basemap around an input geometry.
}
\details{
\strong{Note}: for now it really only works around a polygon
}
\examples{
basemap = osmGet(cambridge_town, tolerance = 1.5, tileserver = "mapbox")
plot(basemap)
# skobbler maps look better with a little purple haze
plot(spTransform(gBoundingPoly(cambridge_town, tolerance = 1.5), osm()), add = T, col = col2hex("lavender", "40"), border = NA)
plot(spTransform(cambridge_town, osm()), add = T, col = col2hex("skyblue", "80"))
}
\keyword{spatial}

