% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSpatial.R
\name{yoverx}
\alias{yoverx}
\title{Height / width ratio}
\usage{
yoverx(geom, osm = T)
}
\arguments{
\item{geom}{an \code{sp} geometry object

Since map objects have varying aspect ratios, saving them to files results in lots of whitespace. This function enables you to specify the output dimensions in the same ratio as the \code{sp} object.}
}
\description{
Quick function for getting the 1/aspect ratio of a geometry
}
\examples{
x = 800 # width in pixels
y = yoverx(basemap)
png(filename, width = x, height = y)
}
